子模块修改，注册之前

+++ admin/father_module_delete.php
@@ -6,6 +6,12 @@ if(!isset($_GET['id'])||!is_numeric($_GET['id'])){
     skip('father_module.php','id参数错误');
 }
 $link = connect();
+$query = "select * from son_module where father_module_id={$_GET['id']}";
+$result = execute($link,$query);
+if (mysqli_num_rows($result)) {
+  skip('father_module.php','父版块下面存在子版块');
+  exit;
+}
 $query = "delete from father_module where id={$_GET['id']}";
 execute($link, $query);
 if (mysqli_affected_rows($link) == 1) {
